Client->SyncClient followup: split up source and test files by classes
Auto-refactor renamed whole Client file 9which also contains GenericClient)
to SyncClient, which is a only minor portion of code&tests.

Now it'd be split up more.
